add font generator

+++ include/hob3lbase/vec.h
@@ -351,7 +351,7 @@
             __typeof__(user)) = cmp; \
         cp_v_bsearch_( \
             (void const *)key, \
-            (cp_v_t*)vec, cp_v_esz_(vec), \
+            (cp_v_t*)(size_t)vec, cp_v_esz_(vec), \
             (int(*)(void const *, void const *, void *))_l_cmp, \
             (void*)(size_t)user); \
     })
